moved instrumentation to diff file
